Last changes for search page.
