+++ TrackList/app/(tabs)/search.tsx
@@ -11,7 +11,7 @@ import {
 import { useRouter } from "expo-router";
 import { Ionicons } from "@expo/vector-icons";
 
-// Example item data
+// Dummy item data
 const data = [
     {
         id: "1",
@@ -19,7 +19,7 @@ const data = [
         artist: "Bad Bunny",
         year: "2025",
         type: "Album",
-        image: "../assets/images/dtmf.png"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "2",
@@ -27,7 +27,7 @@ const data = [
         artist: "Jhayco",
         year: "2024",
         type: "Song",
-        image: "https://t2.genius.com/unsafe/504x504/https%3A%2F%2Fimages.genius.com%2F79fb0f0fe10a96205020ee6aa1c65226.1000x1000x1.png"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "3",
@@ -35,7 +35,7 @@ const data = [
         artist: "Kendrick Lamar",
         year: "2024",
         type: "Song",
-        image: "https://t2.genius.com/unsafe/504x504/https%3A%2F%2Fimages.genius.com%2F95cfea0187b37c7731e11d54b07d2415.1000x1000x1.png"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "4",
@@ -43,7 +43,7 @@ const data = [
         artist: "Billie Eilish",
         year: "2024",
         type: "Song",
-        image: "https://t2.genius.com/unsafe/504x504/https%3A%2F%2Fimages.genius.com%2F5e7bf410789d01a90983b2641b88e5bd.1000x1000x1.png"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "5",
@@ -51,7 +51,7 @@ const data = [
         artist: "LISA",
         year: "2025",
         type: "Song",
-        image: "https://t2.genius.com/unsafe/504x504/https%3A%2F%2Fimages.genius.com%2Fbfa9a5d7d539309ba5f562cf417b2172.1000x1000x1.png"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "6",
@@ -59,7 +59,7 @@ const data = [
         artist: "Doja Cat",
         year: "2024",
         type: "Song",
-        image: "https://t2.genius.com/unsafe/504x504/https%3A%2F%2Fimages.genius.com%2Fa14f94eeaeb85446d1842cd8de8c571d.500x500x1.jpg"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "7",
@@ -67,7 +67,7 @@ const data = [
         artist: "Romeo Santos",
         year: "2022",
         type: "Album",
-        image: "https://t2.genius.com/unsafe/600x600/https%3A%2F%2Fimages.genius.com%2F9c679c5ed84e74f7de85732acf60c38c.1000x1000x1.png"
+        image: require("../../assets/images/jhayco.png"),
     },
     {
         id: "8",
@@ -75,7 +75,7 @@ const data = [
         artist: "Mora",
         year: "2023",
         type: "Song",
-        image: "https://t2.genius.com/unsafe/600x0/https%3A%2F%2Fimages.genius.com%2Ff49b02e184f7be0c388b2b7cde37341a.1000x1000x1.png"
+        image: require("../../assets/images/jhayco.png"),
     }
 ];
 
@@ -90,7 +90,7 @@ export default function Search() {
     // @ts-ignore
     const renderItem = ({ item }) => (
         <View style={styles.itemContainer}>
-            <Image source={{ uri: item.image }} style={styles.itemImage} />
+            <Image source={item.image} style={styles.itemImage} />
             <View style={styles.itemTextContainer}>
                 <View style={{ flexDirection: "row", alignItems: "center" }}>
                     <Text style={styles.itemTitle}>{item.title} </Text>
@@ -105,11 +105,7 @@ export default function Search() {
 
     return (
         <View style={styles.container}>
-            {/* Mimic a nav bar with "Search" title */}
-            <View style={styles.navBar}>
-                <Text style={styles.navBarTitle}>Search</Text>
-            </View>
-
+            <Text style={styles.searchTitle}>Search</Text>
             {/* Search bar */}
             <View style={styles.searchContainer}>
                 <Ionicons name="search-outline" size={20} color="#666" style={styles.searchIcon} />
@@ -132,19 +128,6 @@ export default function Search() {
                 renderItem={renderItem}
                 contentContainerStyle={styles.listContent}
             />
-
-            {/* Bottom tab bar with icons */}
-            <View style={styles.tabBar}>
-                <TouchableOpacity onPress={() => router.push("/menu")}>
-                    <Ionicons name="home-outline" size={32} color="#333" />
-                </TouchableOpacity>
-                <TouchableOpacity onPress={() => router.push("/search")}>
-                    <Ionicons name="search-outline" size={40} color="#FF6600" />
-                </TouchableOpacity>
-                <TouchableOpacity onPress={() => router.push("/profile")}>
-                    <Ionicons name="person-outline" size={32} color="#333" />
-                </TouchableOpacity>
-            </View>
         </View>
     );
 }
@@ -152,21 +135,15 @@ export default function Search() {
 const styles = StyleSheet.create({
     container: {
         flex: 1,
-        backgroundColor: "#D9D9D9"
+        backgroundColor: "#fcfcfc"
     },
-    navBar: {
-        paddingTop: 50, // for iOS status bar offset
-        paddingBottom: 10,
-        backgroundColor: "#fff",
-        alignItems: "center",
-        borderBottomWidth: 1,
-        borderBottomColor: "#eee"
-    },
-    navBarTitle: {
-        paddingTop: 10,
-        fontSize: 18,
-        fontWeight: "600"
-    },
+    searchTitle: {
+        fontSize: 28,
+        fontWeight: 'bold',
+        color: 'black',
+        textAlign: 'center',
+        marginTop: 55,  
+      },
     searchContainer: {
         flexDirection: "row",  // Align items horizontally
         alignItems: "center",  // Center vertically
@@ -227,17 +204,4 @@ const styles = StyleSheet.create({
         fontSize: 14,
         color: "#666"
     },
-    tabBar: {
-        flexDirection: "row",
-        justifyContent: "space-around",
-        alignItems: "center",
-        height: 80,
-        borderTopWidth: 1,
-        borderTopColor: "#ccc",
-        backgroundColor: "#D9D9D9",
-        position: "absolute",
-        bottom: 0,
-        left: 0,
-        right: 0
-    }
 });
